% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge_data.R
\docType{data}
\name{discharge_data}
\alias{discharge_data}
\title{A sample dataset of hospital discharges}
\format{A data frame with 1,200 rows (one for each discharge)
and 7 variables:
\describe{
  \item{drg}{diagnosis related group (drg) code for the admission}
  \item{age}{patient age}
  \item{zip5}{patient 5 digit zip code of residence}
  \item{hosp_id}{hospital identifier}
  \item{hospital}{hospital name}
  \item{sys_id}{system identifier}
  \item{system}{system name}

}}
\source{
Created as part of healthcare.antitrust package.
}
\usage{
discharge_data
}
\description{
A sample dataset of hospital discharges containing patient
characteristics.
}
\keyword{datasets}
