% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare.antitrust-package.R
\docType{package}
\name{healthcare.antitrust-package}
\alias{healthcare.antitrust}
\alias{healthcare.antitrust-package}
\title{healthcare.antitrust: Healthcare Antitrust Analysis}
\description{
Antitrust analysis of healthcare markets. Contains functions to implement the semiparametric estimation technique described in Raval, Rosenbaum, and Tenn (2017) "A Semiparametric Discrete Choice Model: An Application to Hospital Mergers" \doi{10.1111/ecin.12454}.
}
\details{
For more details see the example vignette by typing:
\code{vignette("semipar_example", package = "healthcare.antitrust")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpanhans/healthcare.antitrust}
}

}
\author{
\strong{Maintainer}: Matthew T Panhans \email{mpanhans@gmail.com} (\href{https://orcid.org/0000-0001-7874-6740}{ORCID})

}
\keyword{internal}
