% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_defn.R
\name{cell_defn}
\alias{cell_defn}
\title{Allocate Observations to Cells}
\usage{
cell_defn(D, S_min, layers)
}
\arguments{
\item{D}{Dataset of discharges. The one required variable is adm, which
indicates how many admissions are represented by the observation.}

\item{S_min}{Minimize cell size.}

\item{layers}{A list of lists. Each layer is a list of the variables
on which observations will be allocated to cells. The layers should
be ordered in decreasing refinement, so that observations not allocated
to a cell meeting the minimum size threshold can be allocated by a
coaser layer.}
}
\description{
Take a dataset of hospital discharges, and allocate them to cells that
meet a minimum threshold cell size.
}
\details{
The output is a list of datasets. The first item in the list
is the original dataset, with observations allocated to cells. The
second item in the list is a dataset of unmatched observations.

For more details see the example vignette by typing:
\code{vignette("semipar_example", package = "healthcare.antitrust")}
Make sure you installed the package including the vignettes:
\code{devtools::install_github("mpanhans/healthcare.antitrust")}
}
