% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_defn.R
\name{cell_defn}
\alias{cell_defn}
\title{Allocate Observations to Cells}
\usage{
cell_defn(D, s_min, layers, adm = "adm")
}
\arguments{
\item{D}{Dataset of discharges.}

\item{s_min}{Minimize cell size.}

\item{layers}{A list of lists. Each layer is a list of the variables
on which observations will be allocated to cells. The layers should
be ordered in decreasing refinement, so that observations not allocated
to a cell meeting the minimum size threshold can be allocated by a
more coase layer.}

\item{adm}{Name of variable which indicates the number of admissions
represented by an observation. Default variable name is \code{adm}.}
}
\description{
Take a dataset of hospital discharges, and allocate them to cells that
meet a minimum threshold cell size.
}
\details{
The output is a list of datasets. The first item in the list
is the original dataset, with observations allocated to cells. The
cells are designated by the variable \code{cell}, and the layer number in
which the observation is allocated to the cell is given by the variable
\code{cell_type}. The second item in the output list is a dataset of the
unassigned observations.

If the variable \code{adm} is not supplied, the function will assume that
each observation represents one admission, and a variable \code{adm} will
be created in the output dataset to indicate this.


For more details see the example vignette by typing:
\code{vignette("semipar_example", package = "healthcare.antitrust")}
}
