% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_calc.R
\name{div_calc}
\alias{div_calc}
\title{Diversion Ratio Calculator}
\usage{
div_calc(
  D,
  cell = "cell",
  hosp_id = "hosp_id",
  hospital = "hospital",
  sys_id = "sys_id",
  party_ind = "party_ind",
  adm = "adm",
  dropDegenerateCell = TRUE
)
}
\arguments{
\item{D}{Dataset of hospital discharges, with required variables:
\code{cell}, \code{hosp_id}, \code{hospital}, \code{sys_id},
\code{party_ind}, \code{adm}. Use other function arguments to
indicate alternative variable names to the defaul names.}

\item{cell}{Name of variable specifying cell to which each observation
has been allocated. Default variable name is \code{cell}. Can be
created by \code{cell_defn} function.}

\item{hosp_id}{Name of variable specifying (numeric) hospital
identifier. Default variable name is \code{hosp_id}.}

\item{hospital}{Name of variable specifying (string) hospital name.
Default variable name is \code{hospital}.}

\item{sys_id}{Name of variable specifying (numeric) system identifier.
Default variable name is \code{sys_id}.}

\item{party_ind}{Name of indicator variable for whether hospital is a
merging party from which diversions should be calculated. Default
variable name is \code{party_ind}.}

\item{adm}{Name of variable indicating the number of admissions
represented by the observation. Set = 1 for every row if each
observation represents one admission.}

\item{dropDegenerateCell}{logical; specifies how to treat cells with a
100\% within-system share. If TRUE, observations in degenerate, 100\% share
cells will be ignored in the diversion ratio calculation. If FALSE,
any such individuals will be assigned to the outside option, but still
included in the denominator, so that the inside-option diversion will total
less than 100\%.}
}
\description{
Calculates hospital-level diversion ratios, once cells have been
defined.
}
\details{
Two objects are given as output. The first is a matrix giving
 hospital-level diversions from party hospitals to all other hospitals.
 The second object is a matrix that aggregates party hospitals to
 systems, thus giving diversions from party systems to all other
 hospitals. For system-to-system diversions, set \code{hosp_id} and
 \code{hospital} equal to corresponding system-level identifiers.
 Patients are not allowed to divert to within-system alternative
 hospitals.

For more details see the example vignette by typing:
\code{vignette("semipar_example", package = "healthcare.antitrust")}
}
