% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_calc.R
\name{div_calc}
\alias{div_calc}
\title{Diversion Ratio Calculator}
\usage{
div_calc(D)
}
\arguments{
\item{D}{Dataset of hospital discharges. Required variables:
\itemize{
\item cell: the id of the cell each observation has been allocated
\item hosp_id: hospital identifier (numeric)
\item Hospital: name of hospital (string)
\item sys_id: system identifier (numeric)
\item System: name of system (string)
\item party_ind: indicator for party hospitals
\item adm: the number of observations represented by the observation,
= 1 for all if each observation is one admission
}}
}
\description{
Calculates hospital-level diversion ratios, once cells have been
defined.
}
\details{
The output is hospital-level diversions for party hospitals. For system
level diversion, let hosp_id and Hospital be equal to corresponding
system-level identifiers. Patients are not allowed to divert to
within-system alternative hospitals.

For more details see the example vignette:
\href{../doc/semipar_example.html}{\code{vignette("semipar_example", package = "healthcare.antitrust")}}
}
