% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div_calc.R
\name{div_calc}
\alias{div_calc}
\title{Diversion Ratio Calculator}
\usage{
div_calc(D, dropDegenerateCell = TRUE)
}
\arguments{
\item{D}{Dataset of hospital discharges. Required variables:
\itemize{
\item cell: the id of the cell each observation has been allocated
\item hosp_id: hospital identifier (numeric)
\item hospital: name of hospital (string)
\item sys_id: system identifier (numeric)
\item system: name of system (string)
\item party_ind: indicator for party hospitals
\item adm: the number of observations represented by the observation,
= 1 for all if each observation is one admission
}}

\item{dropDegenerateCell}{logical; specifies how to treat cells with a
100\% within-system share. If TRUE, observations in degenerate, 100\% share
cells with be ignored in the diversion ratio calculation. If FALSE,
any such individuals will be assigned to the outside option, but still
included in the denominator, so that the inside-option diversion will total
less than 100\%.}
}
\description{
Calculates hospital-level diversion ratios, once cells have been
defined.
}
\details{
Two objects are given as output. The first gives hospital-
level diversions from party hospitals to all other hospitals. The
second object is a matrix that aggregates party hospitals to systems,
thus giving diversions from party systems to all other hospitals.
For system-to-system level diversions, set hosp_id and hospital
equal to corresponding system-level identifiers. Patients are not
allowed to divert to within-system alternative hospitals.

For more details see the example vignette by typing:
\code{vignette("semipar_example", package = "healthcare.antitrust")}
}
