% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthcare.antitrust-package.R
\docType{package}
\name{healthcare.antitrust-package}
\alias{healthcare.antitrust}
\alias{healthcare.antitrust-package}
\title{healthcare.antitrust: Tools for Healthcare Antitrust Analysis}
\description{
Tools for antitrust analysis of
    healthcare markets. It contains functions to implement the 
    semiparametric estimation technique described in Raval, Rosenbaum,
    and Tenn (2017) "A Semiparametric Discrete Choice Model: An Application
    to Hospital Mergers" <doi:10.1111/ecin.12454>.
}
\details{
For more details see the example vignette by typing:
\code{vignette("semipar_example", package = "healthcare.antitrust")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpanhans/healthcare.antitrust}
}

}
\keyword{internal}
