% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtp_calc.R
\name{wtp_calc}
\alias{wtp_calc}
\title{Willingness-to-Pay Calculator}
\usage{
wtp_calc(D)
}
\arguments{
\item{D}{Dataset of hospital discharges. Required variables:
\itemize{
\item cell: the id of the cell each observation has been allocated
\item sys_id: system identifier (numeric)
\item party_ind: indicator for party hospitals
\item adm: the number of observations represented by the observation,
= 1 for all if each observation is one admission
\item weight: the designated DRG weight of admission; =1 if
observations should be equally weighted.
}}
}
\description{
This function calculates the system-level Willingness-To-Pay.
}
\details{
This function calculates the system-level WTP. For use in a WTP
 simulation exercise of the "WTP/Q Method" described in Brand and Balan
 (2018).
 Hospital systems need to be numbered by sys_id, with distinct sys_id
 for all independent hospitals
 The functions outputs both weighted and unweighted WTP by system.
 For example, is it common to weight each observation by DRG weight.
 For only unweighted results, weight should be = 1 for all observations,
 then two measures will be equivalent.

For more details see the example vignette by typing:
\code{vignette("semipar_example", package = "healthcare.antitrust")}
}
